chatbot-backend/requirements.txt
fastapi
uvicorn
pydantic
asyncpg
python-dotenv
qdrant-client
openai
google-generativeai
passlib[bcrypt]
python-jose[cryptography]